Multi server setup

+++ cogs/destiny_api_cogs.py
@@ -17,17 +17,21 @@ class destiny_api_cogs(commands.Cog, name='Destiny Commands'):
 
 
     # this command shows a user their current power, highest power level of each equipement piece, and needed power to hit the next level.
-    @commands.command(name = 'power', help = "`~power <steam_name> <class>` Class should be warlock/hunter/titan (not case sensitive).")
-    async def power(self, ctx, steam_name: str, character: str, platform: int = 3, OAuth = True):
-        # get [memberID, membershipType]
-        player_info = await destiny_helpers.get_member_info(steam_name, platform)
+    @commands.command(name = 'power', help = "`~power<class> optional:<steam_name>` Class should be warlock/hunter/titan (not case sensitive).")
+    async def power(self, ctx, character: str, steam_name: str = "", platform: int = 3, OAuth = True):
 
-        if OAuth:
-            # get access token for discordID/memberID combo
-            access_token = await destiny_helpers.get_user_token(ctx.message.author.id, player_info[0])
-            # if we got an error message back, set OAuth to False and continue
-            if (access_token == "refresh token expired" or access_token == "token not found"):
-                OAuth = False
+        if steam_name == "":
+            player_info = await destiny_helpers.get_member_info_Oauth(ctx.message.author.id)
+        else:
+            # get [memberID, membershipType, displayName]
+            player_info = await destiny_helpers.get_member_info(steam_name, platform)
+
+            if OAuth:
+                # get access token for discordID/memberID combo
+                access_token = await destiny_helpers.get_user_token(ctx.message.author.id, player_info[0])
+                # if we got an error message back, set OAuth to False and continue
+                if (access_token == "refresh token expired" or access_token == "token not found"):
+                    OAuth = False
         
         # get player character info [memberID, membershipType, character_class, char_ids, char_id, emblem]
         player_char_info = await destiny_helpers.get_player_char_info(player_info[0], player_info[1], character, OAuth, access_token)
@@ -39,26 +43,30 @@ class destiny_api_cogs(commands.Cog, name='Destiny Commands'):
         high_items = await destiny_helpers.get_max_power_list(items)
 
         # get formatted message string
-        embed = await destiny_helpers.format_power_message(high_items, player_char_info, steam_name)
+        embed = await destiny_helpers.format_power_message(high_items, player_char_info, player_info[2])
 
         # send message to channel
         await ctx.send(embed = embed)
 
-        # delete command message to keep channels clean
-        await ctx.message.delete()
+        # delete command message to keep channels clean if not a dm
+        if not isinstance(ctx.channel, discord.channel.DMChannel):
+            await ctx.message.delete()
 
-    @commands.command(name = 'next_power', help = "`~next_power <steam_name> <class>` Class should be warlock/hunter/titan (not case sensitive).")
-    async def next_power(self, ctx, steam_name: str, character: str, platform: int = 3, OAuth = True):
+    @commands.command(name = 'level', help = "`~level <class> optional:<steam_name>` Class should be warlock/hunter/titan (not case sensitive).")
+    async def level(self, ctx, character: str, steam_name: str = "", platform: int = 3, OAuth = True):
         
-        # get memberID and membershipType
-        player_info = await destiny_helpers.get_member_info(steam_name, platform)
-        
-        if OAuth:
-            # get access token for discordID/memberID combo
-            access_token = await destiny_helpers.get_user_token(ctx.message.author.id, player_info[0])
-            # if we got an error message back, set OAuth to False and continue
-            if (access_token == "refresh token expired" or access_token == "token not found"):
-                OAuth = False
+        if steam_name == "":
+            player_info = await destiny_helpers.get_member_info_Oauth(ctx.message.author.id)
+        else:
+            # get [memberID, membershipType, displayName]
+            player_info = await destiny_helpers.get_member_info(steam_name, platform)
+
+            if OAuth:
+                # get access token for discordID/memberID combo
+                access_token = await destiny_helpers.get_user_token(ctx.message.author.id, player_info[0])
+                # if we got an error message back, set OAuth to False and continue
+                if (access_token == "refresh token expired" or access_token == "token not found"):
+                    OAuth = False
 
         # get player character info
         player_char_info = await destiny_helpers.get_player_char_info(player_info[0], player_info[1], character, OAuth, access_token)
@@ -78,8 +86,9 @@ class destiny_api_cogs(commands.Cog, name='Destiny Commands'):
         # send message to channel
         await ctx.send(embed = embed)
         
-        # delete command message to keep channels clean
-        await ctx.message.delete()
+        # delete command message to keep channels clean if not a dm
+        if not isinstance(ctx.channel, discord.channel.DMChannel):
+            await ctx.message.delete()
 
 
     @commands.command(name = 'reload_manifest', hidden = True)
@@ -88,8 +97,9 @@ class destiny_api_cogs(commands.Cog, name='Destiny Commands'):
          # load manifest
         await destiny_helpers.get_manifest()
 
-        # delete command message to keep channels clean
-        await ctx.message.delete()
+        # delete command message to keep channels clean if not a dm
+        if not isinstance(ctx.channel, discord.channel.DMChannel):
+            await ctx.message.delete()
 
     # this command sends users a url to authenticate with Bungie.net.
     @commands.command(name = 'authenticate', help = "`~authenticate`, Bot will DM you a link to authenticate with Bungie.net")
@@ -101,8 +111,9 @@ class destiny_api_cogs(commands.Cog, name='Destiny Commands'):
         await ctx.message.author.create_dm()
         await ctx.message.author.dm_channel.send(f'Please use the below link to authenticate with Bungie.net.  It may freeze on the final page, please give it time to finish.\n{url}')
 
-        # delete command message to keep channels clean
-        await ctx.message.delete()
+        # delete command message to keep channels clean if not a dm
+        if not isinstance(ctx.channel, discord.channel.DMChannel):
+            await ctx.message.delete()
     
 
 def setup(bot):
